+++ Sources/Muesli/Models/RecordingOverlayAnchor.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+enum RecordingOverlayAnchor: String, CaseIterable, Identifiable {
+    case top
+    case bottom
+
+    var id: String { rawValue }
+
+    var label: String {
+        switch self {
+        case .top:
+            "Top"
+        case .bottom:
+            "Bottom"
+        }
+    }
+}